pyproject.toml
# pyproject.toml

[build-system]
  requires = ["setuptools>=65.0.0"]
  build-backend = "setuptools.build_meta"

[project]
  name = "notion-recurring-tasks"
  version = "1.0.0"
  authors = [
    { name = "Theodor Blom", email = "me@theodorblom.com" },
  ]
  description = "Using the Notion API to achieve recurring tasks within Notion"
  readme = "README.md"
  requires-python = ">=3.12,<3.13"
  classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
  ]
  dependencies = [
    "requests >= 2.31, < 3",
    "python-dateutil >= 2.8.2, < 3",
    "python-dotenv >= 1.0.0, < 2",
  ]

[project.optional-dependencies]
  dev = [
    "pytest >= 7.4.3, < 8",
    "pytest-cov >= 4.1.0, < 5",
    "pytest-mock >= 3.12.0, < 4",
  ]

[tool.pytest.ini_options]
  addopts = "-v --cov='src'"
  testpaths = "tests"

[project.urls]
  homepage = "https://github.com/Theeoi/notion-recurring-tasks"

[project.scripts]
  notion-recurring-tasks = "notion_recurring_tasks.main:main"